Remove verbose output and add barLength parameter
Verbose output was removed to allow the bar element to be invisible when
cmus is not running or outputs an unknown status. This prevents error
clutter in the status bar.
The 'barLength' parameter was added to allow for simpler, explicit
configuration of this value.
The barLength was set to 7 to save space in my bar. However, this value
can be set to anything you desire.